refactor activity report dirs-files structure
