+++ package.json
@@ -40,6 +40,12 @@
     "pg-format": "1.0.4"
   },
   "devDependencies": {
+    "eslint": "^8.28.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.8.0",
     "supertest": "6.1.3",
     "tape": "5.2.2"
   }
