+++ util/activity-report/.eslintrc.json
@@ -1,0 +1,21 @@
+{
+  "env": {
+    "commonjs": true,
+    "es2021": true,
+    "node": true
+  },
+  "extends": ["airbnb-base", "prettier"],
+  "plugins": ["prettier"],
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest"
+  },
+  "rules": {
+    "prettier/prettier": ["error"],
+    "import/extensions": "off",
+    "no-console": "off",
+    "camelcase": "off",
+    "no-restricted-globals": "off",
+    "no-case-declarations": "off"
+  }
+}
