+++ util/activity-report/index.js
@@ -1,0 +1,9 @@
+const { queryAllActivity } = require('./all-activity.js');
+const { queryDeveloperActivity } = require('./developer-activity.js');
+const { queryAvgTxCost } = require('./avg-tx-cost.js');
+
+module.exports = {
+  queryAllActivity,
+  queryDeveloperActivity,
+  queryAvgTxCost,
+};
